+++ src/main/java/main/utils/storage/DB.java
@@ -27,12 +27,12 @@ public class DB {
         }
     }
 
-    public static int setUsefulData(String name, CustomPlayerDataHolder D0) {
+    public static short setUsefulData(String name, CustomPlayerDataHolder D0) {
         try (PreparedStatement statement = connection.prepareStatement("SELECT * FROM data WHERE name = ?")) {
             statement.setString(1, name);
             try (ResultSet resultSet = statement.executeQuery()) {
                 if (resultSet.next()) {
-                    int rank = resultSet.getInt(3);
+                    short rank = resultSet.getShort(3);
                     D0.setRank(rank);
                     D0.setFastCrystals(resultSet.getBoolean(14));
                     return rank;
